Fix CLI, Web, and Test targets to build on Windows (MSVC)
- cli/CMakeLists.txt: use WinHTTP curl stubs and local jsonrpccpp
  sources on MSVC instead of find_package(CURL)
- web/CMakeLists.txt: add Boost 1.82.0 NuGet include path for Beast
  headers; add _WIN32_WINNT definition
- tests/CMakeLists.txt: use C++20 on MSVC (for designated initializers);
  add /Zc:char8_t- to keep u8"" as const char[]; link Windows libs
- tests/Base58Tests.cpp: change uniform_int_distribution<uint8_t> to
  unsigned int (uint8_t not allowed by MSVC STL)
- src/DigiAssetRules.h/.cpp: make operator== const to fix C++20
  ambiguity with gtest's EXPECT_TRUE

+++ tests/Base58Tests.cpp
@@ -36,8 +36,8 @@ TEST(base58, RandomTest) {
 
     std::random_device device;
     std::mt19937 generator(device());
-    std::uniform_int_distribution<uint8_t> d1(1, 100);
-    std::uniform_int_distribution<uint8_t> d2(0, 255);
+    std::uniform_int_distribution<unsigned int> d1(1, 100);
+    std::uniform_int_distribution<unsigned int> d2(0, 255);
 
     auto create_data = [&]() -> std::vector<uint8_t> {
         std::vector<uint8_t> data(d1(generator));
